feat: Refactor SemanticFormField and Introduce Better API for Form-related components (#150)
Using a somewhat compound component-like pattern, I am favoring to use composition for the input control in SemanticFormField instead of a prop that gets converted into a component.

To fix the breaking change:

Instead of using the input prop, provide the element as a child of the SemanticFormField. The SemanticFormField will pass props to this child component for the proper binding and mapping to the label, error, and help message. This is where the compount-like logic is.

BREAKING CHANGE: Use `Input` as a child of `SemanticFormField` instead of the prop.